pyproject.toml
[tool.poetry]
name = "wg-easy-stats-collector"
version = "0.1.0"
description = ""
authors = ["Vladimir Kul'kov <kulkov.vk@dns-shop.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.28.1"
pydantic = "^2.10.5"
greenlet = "^3.1.1"
clickhouse-driver = "0.2.6"
python-dotenv = "^1.0.1"
loguru = "^0.7.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
testcontainers = "^4.9.0"
pytest-asyncio = "^0.25.2"
ruff = "^0.9.2"
polyfactory = "^2.18.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
exclude = ["migrations"]
ignore = ["RUF001", "B008", "RUF002", "RUF003"]
select = [
    "A", "B", "F", "I", "Q", "ASYNC", "N", "W", "UP", "T20", "SIM", "ANN", "PL",
    "PT", "RET", "E", "S", "C4", "EM", "DTZ", "RUF", "TC", "ARG", "TID"
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401", "F403"] # unused imports
"test_*.py" = ["ANN201", "S"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
markers = [
    "no_ch: Tests that doesn't use a database",
]
addopts = ["--strict-markers"]